refactor(host-infra-system): extract config tests into dedicated modules
